# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ArogyaLive API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /facilities/{id}:
    get:
      operationId: getFacility
      tags: [facilities]
      summary: Get facility by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Facility
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Facility"
...
  /footfall:
    get:
      operationId: getFootfall
      tags: [footfall]
      summary: Get footfall records
      parameters:
        - name: facilityId
          in: query
          required: true
          schema:
            type: integer
        - name: days
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Footfall records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FootfallRecord"
  /footfall/log:
    post:
      operationId: logFootfall
      tags: [footfall]
      summary: Log patient footfall
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FootfallInput"
      responses:
        "201":
          description: Logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FootfallRecord"
...
  /doctors:
    get:
      operationId: getDoctors
      tags: [doctors]
      summary: Get doctors for a facility
      parameters:
        - name: facilityId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /attendance:
    get:
      operationId: getAttendance
      tags: [attendance]
      summary: Get attendance for a facility on a date
      parameters:
        - name: facilityId
          in: query
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/mark:
    post:
      operationId: markAttendance
      tags: [attendance]
      summary: Mark doctor attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
...
  /calamity:
    get:
      operationId: getCalamities
      tags: [calamity]
      summary: Get active calamities
      parameters:
        - name: facilityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Calamities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Calamity"
    post:
      operationId: declareCalamity
      tags: [calamity]
      summary: Declare a calamity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalamityInput"
      responses:
        "201":
          description: Declared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calamity"
  /calamity/{id}/resolve:
    post:
      operationId: resolveCalamity
      tags: [calamity]
      summary: Resolve a calamity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Calamity"
...
components:
  schemas:
...
    Facility:
      type: object
      required: [id, name, type, address, latitude, longitude]
      properties:
        id:
          type: integer
        districtId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [PHC, CHC]
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        phone:
          type: ["string", "null"]
        inchargeName:
          type: ["string", "null"]
        totalBedCapacity:
          type: integer
        healthStatus:
          type: ["string", "null"]
          enum: [healthy, watch, critical, null]
        hasActiveCalamity:
          type: boolean
        createdAt:
          type: string
...
    FootfallRecord:
      type: object
      required: [id, facilityId, department, count, logDate]
      properties:
        id:
          type: integer
        facilityId:
          type: integer
        department:
          type: string
        count:
          type: integer
        logDate:
          type: string
        createdAt:
          type: string
    FootfallInput:
      type: object
      required: [facilityId, department, count]
      properties:
        facilityId:
          type: integer
        department:
          type: string
        count:
          type: integer
        logDate:
          type: string
...
    Doctor:
      type: object
      required: [id, facilityId, name, specialization]
      properties:
        id:
          type: integer
        facilityId:
          type: integer
        name:
          type: string
        specialization:
          type: string
        scheduledDays:
          type: string
    AttendanceRecord:
      type: object
      required: [id, doctorId, facilityId, attendanceDate, status]
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        facilityId:
          type: integer
        doctorName:
          type: string
        attendanceDate:
          type: string
        status:
          type: string
          enum: [present, absent, late]
        checkInTime:
          type: ["string", "null"]
    AttendanceInput:
      type: object
      required: [doctorId, facilityId, status]
      properties:
        doctorId:
          type: integer
        facilityId:
          type: integer
        status:
          type: string
          enum: [present, absent, late]
        checkInTime:
          type: string
        attendanceDate:
          type: string
...
    Calamity:
      type: object
      required: [id, facilityId, type, severity, isActive]
      properties:
        id:
          type: integer
        facilityId:
          type: integer
        facilityName:
          type: ["string", "null"]
        type:
          type: string
          enum: [flood, fire, outbreak, infrastructure, other]
        severity:
          type: string
          enum: [low, medium, high]
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        triggeredAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    CalamityInput:
      type: object
      required: [facilityId, type, severity]
      properties:
        facilityId:
          type: integer
        type:
          type: string
          enum: [flood, fire, outbreak, infrastructure, other]
        severity:
          type: string
          enum: [low, medium, high]
        description:
          type: string
...