# lib/api-zod/src/generated/types/calamity.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ArogyaLive API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalamitySeverity } from './calamitySeverity';
import type { CalamityType } from './calamityType';

export interface Calamity {
  id: number;
  facilityId: number;
  /** @nullable */
  facilityName?: string | null;
  type: CalamityType;
  severity: CalamitySeverity;
  /** @nullable */
  description?: string | null;
  isActive: boolean;
  triggeredAt?: string;
  /** @nullable */
  resolvedAt?: string | null;
}