# lib/api-zod/src/generated/types/footfallInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ArogyaLive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FootfallInput {
  facilityId: number;
  department: string;
  count: number;
  logDate?: string;
}